requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests==2.32.0
aiohttp==3.10.11
Werkzeug==3.0.6
Jinja2==3.1.6
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
certifi==2024.7.4
charset-normalizer==3.3.0
idna==3.7
urllib3==2.2.2
asyncio==3.4.3